Fix RoPE shape mismatch and preserve dtype in schedulers
Two critical fixes for Qwen image editing:

1. RoPE shape mismatch: Handle sequence length mismatches in
   _apply_rope_qwen by slicing RoPE embeddings to match actual
   sequence length. Fixes ValueError when mask length differs
   from encoded sequence length by 1 token.

2. Dtype preservation: Ensure all schedulers preserve input dtype
   by casting result back to latents.dtype. Prevents unwanted
   float32 promotion when using bfloat16 weights, which caused
   performance degradation after first iteration.

Also removed debug print statements from qwen_attention.py.

+++ src/mflux/models/common/schedulers/seedvr2_euler_scheduler.py
@@ -55,4 +55,4 @@ class SeedVR2EulerScheduler(BaseScheduler):
             next_sample = (1 - s_norm) * pred_x_0 + s_norm * pred_noise
         else:
             next_sample = pred_x_0
-        return next_sample
+        return next_sample.astype(latents.dtype)
